+++ test/GPU/runtests.jl
@@ -1,0 +1,7 @@
+using CUDA
+using SpeedyWeather
+using Adapt
+using Test
+
+# transforms
+include("spectral_transform.jl")